Better directory handling + cmake/vscode configs
Allows for the explicit definition of the output directory via a '-o'
flag on the command line. The quick implementation sticks the output dir
in the input file ptree, so this has (unintended) side-effect of
allowing configuration of the output directory from the .info file. The
key is post_processing.output_dir.

Additionally, also makes a small change where if you call adamantine
from a different directory than the .info file, it will change the PWD
to that containing directory. This fixes a small bug where adamantine
would just spin its heels waiting on files in the wrong locations.

The added CMakePresets.json can be used via the command by specifying
the preset.

Lastly all of this was in support of decent VSCode integration, so this
adds those configs to be shared. These configs expect that you have the
command-variable plugin installed.